+++ frontend/src/components/GVDBsRetrParams/functions.ts
@@ -1,4 +1,4 @@
-import type { GVDBsDefRetrParams } from "./types";
+import type { GVDBsDefRetrParams, GVDBsShortRetrParams } from "./types";
 
 export function getRetrParamsShortNameFromFullStr(fullStr: string) {
   const d = JSON.parse(fullStr) as GVDBsDefRetrParams;
@@ -16,6 +16,24 @@ export function getRetrParamsShortNameFromFullStr(fullStr: string) {
     text = `MMR: ${sk_mmr.k}/${sk_mmr.fetch_k}/${sk_mmr.lambda_mult}`;
   } else if (search_type === "similarity_score_threshold") {
     text = `SST: ${sk_sst.k}/${sk_sst.score_threshold}`;
+  }
+  return text;
+}
+
+export function getRetrParamsShortNameFromShortStr(shortStr: string) {
+  const d = JSON.parse(shortStr) as GVDBsShortRetrParams;
+  let text = "N/A";
+  let search_type = d.search_type;
+  let sk = d.search_kwargs;
+  if (!(search_type && sk)) {
+    return text;
+  }
+  if (search_type === "similarity") {
+    text = `SIM: ${sk.k}`;
+  } else if (search_type === "mmr") {
+    text = `MMR: ${sk.k}/${sk.fetch_k}/${sk.lambda_mult}`;
+  } else if (search_type === "similarity_score_threshold") {
+    text = `SST: ${sk.k}/${sk.score_threshold}`;
   }
   return text;
 }